improved configuration of Parse.serverURL setting
By setting up the Parse.serverURL configuration locally as something like "http://localhost:XXXX/parse" (where XXXX is the port number Parse will listen to) and on Heroku as "https://yourAppName.herokuapp.com" (where "yourAppName" is your app's name on Heroku). This is why (in #2.2.5) the PARSER_SERVER_URL and/or SERVER_URL must NOT include "/parse" at the end as the PARSE_MOUNT is being appended explicitly.

Also a VERBOSE variable is inserted. Turn on for debugging by uncommenting.